requirements.txt
asgiref==3.3.1
Django==3.1.7
django-admin-rangefilter==0.7.0
django-background-tasks==1.2.5
django-compat==1.0.15
psycopg2==2.8.6
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
gunicorn==20.0.4